requirements.txt
pandas
numpy
matplotlib
scikit-learn
jieba
pdfplumber
python-docx
openpyxl